requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.9
langchain-community==0.0.21
langchain-groq==0.0.1
chromadb==0.4.22
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart==0.0.6
groq==0.4.1
